requirements.txt
Django==3.2.1
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-cacheops==6.0
django-redis==4.12.1
psycopg2-binary==2.8.6
phonenumbers==8.12.22
djangorestframework-jwt==1.11.0
django-environ==0.4.5
coreapi==2.3.3
